+++ src/java/com/echothree/model/control/search/server/graphql/SearchItemsResultObject.java
@@ -35,7 +35,7 @@ public class SearchItemsResultObject
     
     @GraphQLField
     @GraphQLDescription("count")
-    public Integer getCount() {
+    public Long getCount() {
         return result == null ? null : result.getCount();
     }
